feat: Participant 추상화 추가
